Update DG with new UI components Refactor LessonCard and StudentCard into its own package
