+++ docs/diagrams/uidiagrams/InfoPanelClassDiagram.puml
@@ -1,0 +1,14 @@
+@startuml
+!include ../style.puml
+skinparam arrowThickness 1.1
+skinparam arrowColor UI_COLOR_T4
+skinparam classBackgroundColor UI_COLOR
+
+Class "{abstract}\nListPanel" as InfoPanel
+Class "StudentInfoPanel"
+Class "LessonInfoPanel"
+
+LessonInfoPanel --|> InfoPanel
+StudentInfoPanel --|> InfoPanel
+
+@enduml
