+++ docs/diagrams/uidiagrams/ListPanelClassDiagram.puml
@@ -1,0 +1,14 @@
+@startuml
+!include ../style.puml
+skinparam arrowThickness 1.1
+skinparam arrowColor UI_COLOR_T4
+skinparam classBackgroundColor UI_COLOR
+
+Class "{abstract}\nListPanel" as ListPanel
+Class "StudentListPanel"
+Class "LessonListPanel"
+
+LessonListPanel --|> ListPanel
+StudentListPanel --|> ListPanel
+
+@enduml
