+++ src/main/java/seedu/address/ui/card/LessonCard.java
@@ -1,16 +1,15 @@
-package seedu.address.ui;
+package seedu.address.ui.card;
 
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
-import javafx.scene.layout.Region;
 import seedu.address.model.lesson.Lesson;
 
 /**
  * An UI component that displays information of a {@code Student}.
  */
-public class LessonCard extends UiPart<Region> {
+public class LessonCard extends Card {
 
     private static final String FXML = "LessonListCard.fxml";
